refactor(architecture): 重构配置管理模块目录结构
改动说明

- 为什么改：原 apps/backend/configManager.ts 位于根目录不符合项目架构规范，需要将其迁移到 @/lib/config/ 目录以保持代码组织的一致性
- 改了什么：
  - 将 configManager.ts 及其测试文件迁移至 apps/backend/lib/config/ 目录
  - 删除根目录下的旧文件 configManager.ts、configManager.json5.test.ts、configManager.jsonc.test.ts、configManager.test.ts
  - 更新所有引用 @root/configManager.js 的模块为 @/lib/config/configManager.js，共涉及 50+ 文件
  - 导入路径从 import { configManager } from "@root/configManager.js" 更新为 import { configManager } from "@/lib/config/configManager.js"
- 影响范围：
  - 导出接口保持不变，完全向后兼容
  - 所有依赖 ConfigManager 的模块需要更新导入路径
  - 单元测试已同步迁移
- 验证方式：
  - 执行 pnpm type:check 确保类型检查通过
  - 执行 pnpm test 确保所有测试用例通过
  - 配置相关功能（ConfigService、ConfigApiHandler、MCPServiceManager）功能正常
