+++ apps/backend/adapters/__tests__/ConfigAdapter.test.ts
@@ -4,14 +4,14 @@
  */
 
 import { isAbsolute } from "node:path";
-import { MCPTransportType } from "@/lib/mcp/types";
-import { describe, expect, it } from "vitest";
 import type {
   LocalMCPServerConfig,
   MCPServerConfig,
   SSEMCPServerConfig,
   StreamableHTTPMCPServerConfig,
-} from "../../configManager.js";
+} from "@/lib/config/configManager.js";
+import { MCPTransportType } from "@/lib/mcp/types";
+import { describe, expect, it } from "vitest";
 import {
   ConfigValidationError,
   convertLegacyConfigBatch,
