+++ apps/backend/cli.test.ts
@@ -130,7 +130,7 @@ vi.mock("ora", () => ({
   })),
 }));
 
-vi.mock("./configManager", () => ({
+vi.mock("@/lib/config/configManager.js", () => ({
   configManager: {
     configExists: vi.fn(),
     getMcpEndpoint: vi.fn(),
@@ -205,7 +205,7 @@ describe("CLI 命令行工具", () => {
     mockFs = vi.mocked(fs) as FsMock;
     mockOs = vi.mocked(os) as OsMock;
     mockPath = vi.mocked(path) as PathMock;
-    const configManagerModule = await import("./configManager");
+    const configManagerModule = await import("@/lib/config/configManager.js");
     mockConfigManager = vi.mocked(
       configManagerModule.configManager
     ) as ConfigManagerMock;
