+++ apps/backend/handlers/CozeApiHandler.ts
@@ -3,9 +3,9 @@
  * 提供扣子工作空间和工作流相关的 RESTful API 接口
  */
 
+import { configManager } from "@/lib/config/configManager.js";
 import { CozeApiService } from "@/lib/coze";
 import { logger } from "@root/Logger";
-import { configManager } from "@root/configManager";
 import type { CozeWorkflowsParams } from "@root/types/coze";
 import type { Context } from "hono";
