+++ apps/backend/handlers/__tests__/ToolApiHandler.core.test.ts
@@ -3,15 +3,15 @@
  * 测试核心业务逻辑和边界条件处理
  */
 
+import { configManager } from "@/lib/config/configManager.js";
 import type { CozeWorkflow } from "@root/types/coze.js";
 import { ToolType } from "@root/types/toolApi.js";
 import type { Context } from "hono";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { configManager } from "../../configManager.js";
 import { ToolApiHandler } from "../ToolApiHandler.js";
 
 // Mock configManager
-vi.mock("../../configManager.js", () => ({
+vi.mock("@/lib/config/configManager.js", () => ({
   configManager: {
     addCustomMCPTool: vi.fn(),
     getCustomMCPTools: vi.fn(() => []),
