+++ apps/backend/lib/config/configManager.ts
@@ -2,6 +2,7 @@ import { copyFileSync, existsSync, readFileSync, writeFileSync } from "node:fs";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import type { CustomMCPTool as CoreCustomMCPTool } from "@/lib/mcp/types.js";
+import { logger } from "@root/Logger.js";
 import { getEventBus } from "@services/EventBus.js";
 import { TypeFieldNormalizer } from "@utils/TypeFieldNormalizer.js";
 import { validateMcpServerConfig } from "@utils/mcpServerUtils";
@@ -9,7 +10,6 @@ import * as commentJson from "comment-json";
 import dayjs from "dayjs";
 import JSON5 from "json5";
 import * as json5Writer from "json5-writer";
-import { logger } from "./Logger";
 
 // 在 ESM 中，需要从 import.meta.url 获取当前文件目录
 const __dirname = dirname(fileURLToPath(import.meta.url));
