+++ apps/backend/lib/mcp/__tests__/custom.basic.test.ts
@@ -3,7 +3,7 @@
  * 专门测试 Coze 工作流功能
  */
 
-import type { CustomMCPTool } from "@root/configManager.js";
+import type { CustomMCPTool } from "@/lib/config/configManager.js";
 import { getEventBus } from "@root/services/EventBus.js";
 import type { EventBus } from "@services/EventBus.js";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
@@ -21,7 +21,7 @@ vi.mock("@root/Logger.js", () => ({
 }));
 
 // Mock configManager
-vi.mock("@root/configManager.js", () => ({
+vi.mock("@/lib/config/configManager.js", () => ({
   configManager: {
     getCustomMCPTools: vi.fn(),
     getCustomMCPConfig: vi.fn(),
