+++ apps/backend/lib/mcp/transports/__tests__/TransportAdapters.test.ts
@@ -3,12 +3,12 @@
  * 验证 TransportAdapter 和 StdioAdapter 的功能
  */
 
+import type { AppConfig } from "@/lib/config/configManager.js";
+import { configManager } from "@/lib/config/configManager.js";
 import { MCPServiceManager } from "@/lib/mcp/manager.js";
 import { MCPMessageHandler } from "@/lib/mcp/message.js";
-import type { AppConfig } from "@root/configManager.js";
 import type { MCPMessage, MCPResponse } from "@root/types/mcp.js";
 import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
-import { configManager } from "../../../../configManager.js";
 import { StdioAdapter } from "../StdioAdapter.js";
 import { ConnectionState, TransportAdapter } from "../TransportAdapter.js";
