+++ apps/backend/middlewares/endpoints.middleware.ts
@@ -3,9 +3,9 @@
  * 负责创建和管理 MCPEndpointApiHandler 实例
  */
 
+import { configManager } from "@/lib/config/configManager.js";
 import type { EndpointManager } from "@/lib/endpoint/index.js";
 import { MCPEndpointApiHandler } from "@handlers/MCPEndpointApiHandler.js";
-import { configManager } from "@root/configManager.js";
 import type { MiddlewareHandler } from "hono";
 import type { AppContext } from "../types/hono.context.js";
