+++ apps/backend/services/ToolCallService.ts
@@ -3,10 +3,10 @@
  * 负责处理手动工具调用的核心逻辑
  */
 
+import { configManager } from "@/lib/config/configManager.js";
 import { ProcessManagerImpl } from "@cli/services/ProcessManager.js";
 import type { Logger } from "@root/Logger.js";
 import { logger } from "@root/Logger.js";
-import { configManager } from "@root/configManager.js";
 
 // 工具调用结果接口
 export interface ToolCallResult {
