+++ apps/backend/services/__tests__/EnvironmentVariables.integration.test.ts
@@ -3,10 +3,10 @@
  * 验证 MCP 服务环境变量传递修复是否正确工作
  */
 
+import type { LocalMCPServerConfig } from "@/lib/config/configManager.js";
 import { MCPTransportType } from "@/lib/mcp";
 import { convertLegacyToNew } from "@adapters/ConfigAdapter.js";
 import { describe, expect, it } from "vitest";
-import type { LocalMCPServerConfig } from "../../configManager.js";
 
 describe("环境变量传递集成测试", () => {
   describe("配置转换", () => {
