+++ apps/backend/services/__tests__/MultiProtocolIntegration.test.ts
@@ -1,3 +1,4 @@
+import { configManager } from "@/lib/config/configManager.js";
 import { MCPServiceManager } from "@/lib/mcp";
 import { MCPTransportType } from "@/lib/mcp";
 import { TransportFactory } from "@/lib/mcp/transport-factory.js";
@@ -5,7 +6,6 @@ import type { MCPServiceConfig } from "@/lib/mcp/types.js";
 import type { Tool } from "@modelcontextprotocol/sdk/types.js";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { setupCommonMocks } from "../../__tests__/index.js";
-import { configManager } from "../../configManager.js";
 
 // Mock dependencies
 vi.mock("@/lib/mcp", () => {
